+++ assets/scss/styles.scss
@@ -98,6 +98,10 @@ html {
   width: 70% !important;
 }
 
+.w-80 {
+  width: 80% !important;
+}
+
 //Alignment
 .justify-content-center {
   justify-content: center;
@@ -2650,6 +2654,26 @@ body {
     width: 40% !important;
   }
 
+  .w-md-50 {
+    width: 50% !important;
+  }
+
+  .w-md-60 {
+    width: 60% !important;
+  }
+
+  .w-md-70 {
+    width: 70% !important;
+  }
+
+  .w-md-80 {
+    width: 80% !important;
+  }
+
+  .w-md-100 {
+    width: 100% !important;
+  }
+
   .content-item {
     min-height: 16rem !important;
   }
@@ -4574,6 +4598,6 @@ body {
   background-color: #f5f5f5;
   color: rgba(0, 0, 0, 0.87);
 }
-.break-word{
+.break-word {
   word-break: break-word;
 }
